根据 review 结果，修改部分代码
