+++ src/components/checkbox/_checkbox.scss
